add authentication

+++ demos/backend/src/database/deleteOrganizationEventProposal.ts
@@ -1,9 +1,30 @@
+import { getUserFromSession } from "./getUserFromSession";
 import { prisma } from "./prisma";
 
-export async function deleteOrganizationEventProposal(eventProposalId: string) {
+export async function deleteOrganizationEventProposal(
+  session: any,
+  eventProposalId: string
+) {
+  const currentUser = await getUserFromSession(session);
+
   try {
-    // TODO validate access
-    return await prisma.eventProposal.delete({
+    const eventProposal = await prisma.eventProposal.findUnique({
+      where: { id: eventProposalId },
+      include: {
+        organization: {
+          select: {
+            members: {
+              where: { publicSigningKey: currentUser.publicSigningKey },
+            },
+          },
+        },
+      },
+    });
+    if (eventProposal.organization.members.length === 0) {
+      throw new Error("Failed");
+    }
+
+    await prisma.eventProposal.delete({
       where: { id: eventProposalId },
     });
   } catch (err) {
